+++ gradle-plugin/src/functionalTest/kotlin/dev/zacsweers/metro/gradle/MetroOptionOverrides.kt
@@ -1,5 +1,5 @@
 // Copyright (C) 2025 Zac Sweers
 // SPDX-License-Identifier: Apache-2.0
-package dev.zacsweers.metro.gradle.incremental
+package dev.zacsweers.metro.gradle
 
 data class MetroOptionOverrides(val enableFullBindingGraphValidation: Boolean? = null)