+++ gradle-plugin/src/functionalTest/kotlin/dev/zacsweers/metro/gradle/MetroProject.kt
@@ -1,6 +1,6 @@
 // Copyright (C) 2025 Zac Sweers
 // SPDX-License-Identifier: Apache-2.0
-package dev.zacsweers.metro.gradle.incremental
+package dev.zacsweers.metro.gradle
 
 import com.autonomousapps.kit.AbstractGradleProject
 import com.autonomousapps.kit.GradleProject